+++ indexer/scale_index.hpp
@@ -49,7 +49,7 @@ public:
 
     ReaderSource<ReaderT> source(reader);
     VarSerialVectorReader<ReaderT> treesReader(source);
-    for (size_t i = 0; i < treesReader.Size(); ++i)
+    for (int i = 0; i < treesReader.Size(); ++i)
       m_IndexForScale.push_back(factory.CreateIndex(treesReader.SubReader(i)));
   }
